requirements.txt
openai
python-dotenv
langgraph
langchain-core
langchain-community
langchain-text-splitters
pandas
scikit-learn
kaggle

